Added support for UserGame layer and connected it with cartPage (#13)

+++ SteamStore/Models/User.cs
@@ -8,8 +8,8 @@ public class User {
     public int UserId { get; set; }
     public string Name { get; set; }
     public string Email { get; set; }
-    public int WalletBalance { get; set; }
-    public int PointsBalance { get; set; }
+    public float WalletBalance { get; set; }
+    public float PointsBalance { get; set; }
     public enum Role { Developer, User }
 
     public Role UserRole { get; set; }
